work for today, class cleanup + new daily command, old one tmp renamed

+++ cogs/listen.py
@@ -75,6 +75,10 @@ class Listen(commands.Cog):
         if not world:
             return
 
+        pre = self.bot.config.get_prefix(message.guild.id)
+        if message.content.lower().startswith(pre.lower()):
+            return
+
         # Report Converter
         if message.content.__contains__("public_report"):
             file = await self.fetch_report(message.content)
@@ -93,10 +97,6 @@ class Listen(commands.Cog):
         if result:
 
             result = set(result)
-            pre = self.bot.config.get_prefix(message.guild.id)
-            if message.content.lower().startswith(pre.lower()):
-                return
-
             coords = [obj.replace('|', '') for obj in result]
             villages = await self.bot.fetch_bulk(world, coords, "village", name=True)
             player_ids = [obj.player_id for obj in villages]
@@ -110,9 +110,8 @@ class Listen(commands.Cog):
                 else:
                     owner = "[Barbarendorf]"
 
-                coord = f"{vil.x}|{vil.y}"
-                good.append(f"[{coord}]({vil.ingame_url}) {owner}")
-                result.remove(coord)
+                good.append(f"{vil.mention} {owner}")
+                result.remove(f"{vil.x}|{vil.y}")
 
             found = '\n'.join(good)
             lost = ','.join(result)
@@ -129,19 +128,26 @@ class Listen(commands.Cog):
                 return
 
         # DS Player/Tribe Converter
-        names = re.findall(r'<(.*?)>', message.clean_content)
-        emojis = re.findall(r'<a?:[a-zA-Z0-9_]+:([0-9]+)>', message.content)
+        if "|" not in message.content:
+            return
 
-        for idc in emojis:
-            for name in names.copy():
-                if idc in name:
-                    names.remove(name)
+        content = message.clean_content
+        mentions = message.mentions.copy()
+        mentions.extend(message.role_mentions)
+        mentions.extend(message.channel_mentions)
 
+        for mention in mentions:
+            if isinstance(mention, discord.Member):
+                raw = f"@{mention.display_name}"
+            else:
+                raw = f"@{mention.name}"
+            content = content.replace(raw, "")
+
+        names = re.findall(r'(?<!\|)\|([\w][^|]*?)\|(?!\|)', message.clean_content)
         if names:
-            names = names[:10]
             parsed_msg = message.clean_content.replace("`", "")
-            ds_objects = await self.bot.fetch_bulk(world, names, name=True)
-            cache = await self.bot.fetch_bulk(world, names, 1, name=True)
+            ds_objects = await self.bot.fetch_bulk(world, names[:10], name=True)
+            cache = await self.bot.fetch_bulk(world, names[:10], 1, name=True)
             ds_objects.extend(cache)
 
             found_names = {}
@@ -153,12 +159,11 @@ class Listen(commands.Cog):
             for name in names:
                 dsobj = found_names.get(name.lower())
                 if not dsobj:
-                    parsed_msg = parsed_msg.replace(f"<{name}>", "[Unknown]")
-                    continue
+                    failed = f"**{name}**<:failed:708982292630077450>"
+                    parsed_msg = parsed_msg.replace(f"|{name}|", failed)
 
-                correct_name = dsobj.name if dsobj.alone else dsobj.tag
-                hyperlink = f"[{correct_name}]({dsobj.ingame_url})"
-                parsed_msg = parsed_msg.replace(f"<{name}>", hyperlink)
+                else:
+                    parsed_msg = parsed_msg.replace(f"|{name}|", dsobj.mention)
 
             current = message.created_at + timedelta(hours=1)
             time = current.strftime("%H:%M Uhr")
@@ -167,7 +172,8 @@ class Listen(commands.Cog):
             embed.set_author(name=title, icon_url=message.author.avatar_url)
             try:
                 await message.channel.send(embed=embed)
-                await message.delete()
+                if not mentions:
+                    await message.delete()
             except discord.Forbidden:
                 pass
